chore: moved duplicaates file to service folder
