+++ packages/cli/test/unit/services/processComparisonFile.test.ts
@@ -60,7 +60,7 @@ import fs from 'fs';
 import { processComparisonFile } from '../../../src/services/processComparisonFile.js';
 import { applyFixes } from '../../../src/services/fixEnv.js';
 import { parseEnvFile } from '../../../src/services/parseEnvFile.js';
-import { findDuplicateKeys } from '../../../src/core/duplicates.js';
+import { findDuplicateKeys } from '../../../src/services/duplicates.js';
 import { resolveFromCwd } from '../../../src/core/helpers/resolveFromCwd.js';
 
 describe('processComparisonFile', () => {
